Move bfConnector stuff to its own file.
This also moves some common http stuff to http.go.
